+++ lib/aruba/setup.rb
@@ -37,6 +37,7 @@ module Aruba
           runtime.announcer.announce :command, event.entity.commandline
           runtime.announcer.announce :timeout, 'exit', event.entity.exit_timeout
           runtime.announcer.announce :timeout, 'io wait', event.entity.io_wait_timeout
+          runtime.announcer.announce :wait_time, 'startup wait time', event.entity.startup_wait_time
           runtime.announcer.announce :full_environment, event.entity.environment
         end
       )
@@ -52,8 +53,10 @@ module Aruba
       runtime.event_bus.register(
         :command_stopped,
         proc do |event|
-          runtime.announcer.announce :stdout, event.entity.stdout
-          runtime.announcer.announce :stderr, event.entity.stderr
+          runtime.announcer.announce(:stdout) { event.entity.stdout }
+          runtime.announcer.announce(:stderr) { event.entity.stderr }
+          runtime.announcer.announce(:command_content) { event.entity.content }
+          runtime.announcer.announce(:command_filesystem_status) { event.entity.filesystem_status }
         end
       )